% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianguletter.R
\name{trianguletter}
\alias{trianguletter}
\title{A right sided triangle of alphabetic letters}
\usage{
trianguletter(x)
}
\arguments{
\item{x}{A number.}
}
\value{
Outputs to console. NULL object returned.
}
\description{
A right sided triangle of alphabetic letters
}
\details{
This function takes a number as input and outputs an increasing
    number of alphabetic letters on top of eachother, resembling
    a right sided triangle.
}
\section{Warning}{

Not tested for numbers over 26!
}

\examples{
trianguletter(10)
}
